Implementado o CRUD completo com funcionalidade de editar aluno

+++ src/app/pages/home/home.component.ts
@@ -2,10 +2,10 @@ import { Component, OnInit } from '@angular/core';
 import { Alunos } from '../../core/services/types/types';
 import { AlunosService } from '../../core/services/alunos.service';
 import { Router } from '@angular/router';
-
+import { RouterModule } from '@angular/router';
 @Component({
   selector: 'app-home',
-  imports: [],
+  imports: [RouterModule],
   templateUrl: './home.component.html',
   styleUrl: './home.component.css'
 })
@@ -19,13 +19,19 @@ export class HomeComponent implements OnInit {
     }); 
   } 
 
-  excluirAluno(id: string): void {
+  excluirAluno(id: string | undefined): void {
+    if (!id) {
+      alert('ID do aluno não encontrado.');
+      return;
+    }
+  
     if (confirm('Tem certeza que deseja excluir este aluno?')) {
       this.service.deleteAluno(id).subscribe(() => {
         this.listaAlunos = this.listaAlunos.filter(a => a.id !== id);
       });
     }
   }
+
